+++ src/App.js
@@ -22,8 +22,8 @@ function App() {
   const [verbs, setVerbs] = useState([]);
   const [deckVerbs, setDeckVerbs] = useState([])
   const [showFlashcards, setShowFlashcards] = useState(false)
-  const [flashcard, setFlashcard] = useState(null)
-  const [flashcardNumber, setFlashcardNumber] = useState(0)
+  const [flashcards, setFlashcards] = useState([])
+
 
   useEffect(() => {
     if (!isLoading & isAuthenticated) {
@@ -64,10 +64,12 @@ function App() {
         const DeckVerbsFromServer = await fetchDeckVerbs(currentDeck.deckId)
         if(DeckVerbsFromServer == null) {
           setDeckVerbs([])
+          setFlashcards([])
         }
         else {
           shuffleArray(DeckVerbsFromServer)
           setDeckVerbs(DeckVerbsFromServer)
+          setFlashcards(DeckVerbsFromServer)
         }
         
       }
@@ -76,28 +78,6 @@ function App() {
     }
   }, [currentDeck])
 
-  useEffect(() => {
-    if(deckVerbs.length > 0 && deckVerbs.length > flashcardNumber) {
-      const getFlashcard = async (deckVerb) => {
-        const phraseFromServer = await fetchPhrase(deckVerb.verbId)
-        const verb = deckVerb.verb
-        const verbId = deckVerb.verbId
-        if(phraseFromServer == null) {
-          const phrase = "no phrase available for this verb yet"
-          const phraseId = "phraseFromServer.phraseId"
-          const flashcard = { phrase, verb, verbId, phraseId }
-          setFlashcard(flashcard)
-        }
-        else {
-          const phrase = phraseFromServer.phrase
-          const phraseId = phraseFromServer.phraseId
-          const flashcard = { phrase, verb, verbId, phraseId }
-          setFlashcard(flashcard)
-        }
-    }
-    getFlashcard(deckVerbs[flashcardNumber])
-  }},[deckVerbs, flashcardNumber])
-
   function shuffleArray(array) {
     for (let i = array.length - 1; i > 0; i--) {
         const j = Math.floor(Math.random() * (i + 1));
@@ -105,7 +85,6 @@ function App() {
     }
 }
 
-  
   //google-oauth2|109641767784145272988
   // fetch decks
   const fetchDecks = async (userUrl) => {
@@ -129,14 +108,7 @@ function App() {
     return data
   }
 
-  const fetchPhrase = async (verbId) => {
-    const url = `https://localhost:44386/api/Vocabulazy/phrases?verbId=${verbId}`
-    const res = await fetch(url)
-    const data = await res.json()
-    const phrase = data[Math.floor(Math.random() * data.length)];
-    
-    return phrase
-  }
+
 
 
 
@@ -212,7 +184,7 @@ function App() {
         {currentDeck && showChangeDeck && <Decks decks={decks} isLoading={isLoading} isAuthenticated={isAuthenticated} setCurrentDeck={setCurrentDeck} currentDeck={currentDeck} />}
         {isAuthenticated && currentDeck && !showAddVerbs && <Deck showFlashcards = {() => setShowFlashcards(true)} deck={currentDeck} setCurrentDeck={setCurrentDeck} verbs={verbs} deckVerbs={deckVerbs} showAddVerbs={showAddVerbs} setShowAddVerbs={setShowAddVerbs} />}
         {deckVerbs && showAddVerbs && <AddVerbs updateDeckVerbs={updateDeckVerbs} verbs={verbs} deckVerbs={deckVerbs} setDeckVerbs={setDeckVerbs} deckId={currentDeck.deckId} deckName={currentDeck.deckName} setShowAddVerbs={setShowAddVerbs} />}
-        {currentDeck && showFlashcards && <Flashcards setFlashcard={setFlashcard} hideFlashcards = {() => setShowFlashcards(false)} deckVerbs = {deckVerbs} flashcard={flashcard} flashcardNumber={flashcardNumber} setFlashcardNumber={setFlashcardNumber} />}
+        {currentDeck && showFlashcards && <Flashcards flashcards={flashcards} setFlashcards={setFlashcards} hideFlashcards = {() => setShowFlashcards(false)} deckVerbs = {deckVerbs} />}
 
       </div>
     </>
